Began Working on Homepage
Edited the index.html file and adjusted the .css that affects it and everything under the Piracy Catalog. I also removed any link that traced back to the old website format. It is no longer accessible through links when visiting the website. Although it is still accessible via typing the URL since the files are still in Github.